+++ public/rss.xml
@@ -4,7 +4,7 @@
         <title>PangLAN's Blog</title>
         <link>https://yinn.party</link>
         <description>PangLAN's Blog</description>
-        <lastBuildDate>Tue, 05 Dec 2023 05:05:37 GMT</lastBuildDate>
+        <lastBuildDate>Sat, 09 Dec 2023 09:34:01 GMT</lastBuildDate>
         <docs>https://validator.w3.org/feed/docs/rss2.html</docs>
         <generator>https://github.com/jpmonette/feed</generator>
         <copyright>All rights reserved 2023,@PangLAN</copyright>
