+++ src/app/layout.tsx
@@ -18,12 +18,16 @@ export default function RootLayout({
     <html lang="en">
       <SessionProvider>
         <body
-          className={`antialiased min-h-screen bg-slate-100 dark:bg-slate-900 text-slate-700 dark:text-slate-300 ${font.className}`}
+          className={`min-h-screen bg-slate-100 text-slate-700 antialiased dark:bg-slate-900 dark:text-slate-400 ${font.className}`}
         >
           <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
-            <div className="max-w-3xl mx-auto py-10 px-4">
-              <Headers />
-              <main>{children}</main>
+            <div className="mx-4 mb-40 mt-8 flex max-w-5xl flex-col md:mt-20 md:flex-row lg:mx-auto lg:mt-32">
+              <aside className="-mx-4 md:mx-0 md:w-[200px] md:flex-shrink-0 md:px-0">
+                <Headers />
+              </aside>
+              <main className="mt-6 flex min-w-0 flex-auto flex-col md:mt-0">
+                {children}
+              </main>
             </div>
             <Footers />
             <Analytics />
